+++ packages/at_lookup/lib/src/at_lookup_impl.dart
@@ -34,6 +34,7 @@ class AtLookupImpl implements AtLookUp {
 
   late int _rootPort;
 
+  @Deprecated("privateKey reference is no longer used")
   String? privateKey;
 
   String? cramSecret;
@@ -573,14 +574,11 @@ class AtLookupImpl implements AtLookUp {
         if (_atChops != null) {
           logger.finer('calling pkam using atchops');
           await pkamAuthenticate(enrollmentId: enrollmentId);
-        } else if (privateKey != null) {
-          logger.finer('calling pkam without atchops');
-          await authenticate(privateKey);
         } else if (cramSecret != null) {
           await cramAuthenticate(cramSecret!);
         } else {
           throw UnAuthenticatedException(
-              'Unable to perform atLookup auth. Private key/cram secret is not set');
+              'Unable to perform atLookup auth. atChops object is not set');
         }
       }
       try {